Refactor event handling to fix hidden errors
It turns out that even though Juju claims that it is attaching resources
when you deploy a bundle referring to store charms w/ resources along
with an overlay that points to local copies of the charms, it actually
doesn't. So the resources were failing, but because of how the handlers
were working, the status was not being surfaced properly and the charm
seemed like it was functional when it wasn't.