Use pipenv to lock requirements & cleanup tox config
This uses pipenv to lock the requirements, with an update target for
updating the dependencies. It also switches the operator framework to
pull from GitHub rather than PyPI to pick up some newer, as yet
unreleased fixes (specifically the config default value handling in the
test harness). Finally, it refactors the tox config to use pipenv,
remove the unused build env, and makes running tox in the repo root
actually run the tests for the charms.