Fix resource failure, change requirements handling, and refactor test config (#10)
* Refactor event handling to fix hidden errors

It turns out that even though Juju claims that it is attaching resources
when you deploy a bundle referring to store charms w/ resources along
with an overlay that points to local copies of the charms, it actually
doesn't. So the resources were failing, but because of how the handlers
were working, the status was not being surfaced properly and the charm
seemed like it was functional when it wasn't.

* Use pipenv to lock requirements & cleanup tox config

This uses pipenv to lock the requirements, with an update target for
updating the dependencies. It also switches the operator framework to
pull from GitHub rather than PyPI to pick up some newer, as yet
unreleased fixes (specifically the config default value handling in the
test harness). Finally, it refactors the tox config to use pipenv,
remove the unused build env, and makes running tox in the repo root
actually run the tests for the charms.

* Ensure build failure is reported well

* Try to cat the pip log

* Try another location for the pip log

* Fix bootstrap installs and remove attempt to cat pip log that does not exist

* Update dependencies to use newly released 0.10.0 ops library from PyPI instead of github, and fix update script

* Some changes per review

* Fix missed stored attribute change

* Fix charm build failing due to global gitconfig

GH automatically includes the git-lfs plugin and configures it in
/etc/gitconfig.  However, the confinement of the charmcraft snap means
that it can see that this file exists but cannot read it, even if the
file permissions should allow it; this breaks git usage within the snap.
To get around this, we move it from the global gitconfig to the user's
.gitconfig file.

* Increase func test job timeout

Sometimes the microk8s setup takes most of the 10 minutes and times out
the rest of the job.