Convert tests to actions-operator and pytest-operator (#18)
* Convert tests to actions-operator and pytest-operator

Clean up the integration test workflow by leveraging actions-operator
for the environment setup and pytest-operator for model helpers and
doing the testing in Python.

* Refactor RBAC management in tests to Workflow

Managing the RBAC feature within the test ran into issues due to the API
server restarting which interrupted the model connection to the
controller. Additionally, it made the tests coupled to microk8s rather
than being able to run against any K8s cluster.

* Fix workflow syntax error

* Fix perm error on snap alias

* Fix env syntax and sg command for disabling RBAC in workflow

* Fix sg syntax in workflow and custom arg passing through tox

* Fix parallel charmcraft builds

Charmcraft checks for and, if not present, adds a LXC remote for it's
build environment image. If multiple builds are started in parallel,
there's a race condition where they both see that the remote isn't
present and then both try to add it, leading to one of them failing.

* More charmcraft pre-init in workflow

* Do builds sequentially to try to get past build issues

* Remove git dependency in requirements

The oci-image library was released to PyPI some time ago, so use that
instead of the git repo.

* Refactor RBAC check to not use wait_for_idle

The behavior of `wait_for_idle` returning immediately as soon as *any*
unit is in error, even with `raise_on_error=False` was causing some
strangeness with the test attempting to resolve a unit before it was in
error (somehow).

* Add some debug logging

* Increase workflow timeout

* Restore default wait_period to try to figure out why it was failing

* Some debugging for inconsistent test behavior